Initial Commit, Reafactor and Rewrite
The original RoboMigrator GUI application used the (currently) main branch of RoboSharp. With this branch (which hopefully replaces main) we will be using code from dev branch of RoboSharp which implemnts the RoboQueue code to simplify the running of multi-job robocopy scenarios.

Currently this branch of RoboMigratorGUI is not working.